try other exec options
