+++ test/napi/perm.js
@@ -1,0 +1,11 @@
+'use strict'
+
+const { execSync } = require('child_process')
+
+const cwd = __dirname
+const stdio = ['inherit', 'inherit', 'inherit']
+const uid = process.getuid()
+const gid = process.getgid()
+const opts = { cwd, stdio, uid, gid }
+
+execSync('touch foo.txt', opts)
